FIX: MACOS Window Title bar bug with window frame size
