+++ src/bonus/hud/hud_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   hud_bonus.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lmells <lmells@student.42.fr>              +#+  +:+       +#+        */
+/*   By: lmells <lmells@student.42adel.org.au>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 12:21:46 by lmells            #+#    #+#             */
-/*   Updated: 2023/12/14 15:59:34 by lmells           ###   ########.fr       */
+/*   Updated: 2023/12/19 10:11:35 by lmells           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
